backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
pypdf
sentence-transformers
faiss-cpu
python-multipart
python-dotenv
tiktoken
